+++ Sources/PixelTools/PixelTools.swift
@@ -1,5 +1,3 @@
-import PixelCore
-
 public enum PixelTools {
-    public static let version = "0.0.0"
+    public static let version = "0.1.0"
 }